pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel~=0.40.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "howso-engine"
dynamic = ["version"]
readme = "README.md"
description = 'Howso Engine and Scikit Estimator for the interpretable Machine Learning and Artificial Intelligence API.'
authors = [
    {name = "Howso Incorporated"},
    {email = "support@howso.com"},
]
requires-python = ">=3.8"
keywords = ["machine", "learning", "artificial", "intelligence"]
license = {file = "LICENSE.txt"}
classifiers = [
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Science/Research",
        "Topic :: Scientific/Engineering :: Artificial Intelligence",
        "License :: Other/Proprietary License",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Operating System :: MacOS",
        "Operating System :: MacOS :: MacOS X",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: Microsoft :: Windows :: Windows 10",
        "Operating System :: Microsoft :: Windows :: Windows 11",
        "Operating System :: POSIX :: Linux",
]
dependencies = [
   "certifi",
   "cryptography>=2.9.2",
   "deprecation>=2.1.0,<3",
   "Faker>=4.1.1",
   "humanize>=4.0.0",
   "mmh3",
   "numpy",
   "pandas",
   "pyjwt>=2.6.0,<3",
   "pyyaml>=6.0.1,<7.0",
   "rich>=12.5.1",
   "scikit-learn",
   "semantic-version>=2.8.5,<3",
   "typing-extensions>=4.1.0,<5.0",
   "urllib3>=1.26.2,<=2",
   "howso-openapi-client==24.0.16",
   "amalgam-lang==2.5.0",
]

[project.optional-dependencies]
dev = [
   "dill>=0.3.4,<1.0",
   "flake8-docstrings",
   "flake8-import-order",
   "flake8>=3.7,<4.0",
   "pip-tools",
   "pipdeptree",
   "pmlb",
   "pycodestyle",
   "pylint",
   "pytest",
   "pytest-cov",
   "pytest-mock>=1.13.0,<1.14.0",
   "urllib3<2.0.0",
]
no-telemetry = [
    "howso-engine-no-telemetry",
]

[project.urls]
homepage = "https://howso.com"
documentation = "https://docs.howso.com/"
repository = "https://github.com/howsoai/howso-engine-py"

[project.scripts]
verify_howso_install = "howso.utilities.installation_verification:main"

[tool.setuptools]
packages = ["howso"]

[tool.isort]
profile = "google"
known_first_party = ["howso"]
line_length = 119
group_by_package = true
use_parenthesis = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 2

[tool.yapf]
based_on_style = "pep8"
column_limit = 119
dedent_closing_brackets = true
split_penalty_after_opening_bracket = 100
split_penalty_for_added_line_split = 300

# version is managed by setuptools_scm
[tool.setuptools_scm]
